+++ JECS/Parsing Logic/NodeManager.cs
@@ -8,6 +8,8 @@ namespace JECS.ParsingLogic
     /// </summary>
     internal static class NodeManager
     {
+        const string KEY_CONCRETE_TYPE = "{JECS_INTERNAL_CONCRETE_TYPE}";
+        
         internal static void SetNodeData<T>(Node node, T data, FileStyle style) => SetNodeData(node, data, typeof(T), style);
         internal static void SetNodeData(Node node, object data, Type type, FileStyle style)
         {
@@ -49,16 +51,33 @@ namespace JECS.ParsingLogic
             // of the file for something temporary.
             string dataAsString = data as string;
             if (type == typeof(string) && (dataAsString.ContainsNewLine() || node.ChildNodes.Count > 0))
+            {
                 MultiLineStringSpecialCaseHandler.SetStringSpecialCase(node, dataAsString, style);
+                return;
+            }
 
-            else if (BaseTypesManager.IsBaseType(type))
+            if (BaseTypesManager.IsBaseType(type))
+            {
                 SetBaseTypeNode(node, data, type, style);
+                return;
+            }
 
-            else if (CollectionTypesManager.IsSupportedType(type))
+            if (CollectionTypesManager.IsSupportedType(type))
+            {
                 CollectionTypesManager.SetCollectionNode(node, data, type, style);
+                return;
+            }
 
-            else
-                ComplexTypes.SetComplexNode(node, data, type, style);
+            if (type.IsAbstract || type.IsInterface)
+            {
+                var concreteTypeNode = node.GetChildAddressedByName(KEY_CONCRETE_TYPE);
+                var concreteType = data.GetType();
+                SetNodeData(concreteTypeNode, concreteType, typeof(Type), style);
+                SetNodeData(node, data, concreteType, style);
+                return;
+            }
+
+            ComplexTypes.SetComplexNode(node, data, type, style);
         }
 
         internal static T GetNodeData<T>(Node node) => (T)GetNodeData(node, typeof(T));
@@ -88,6 +107,13 @@ namespace JECS.ParsingLogic
 
             if (node.HasValue)
                 return RetrieveDataWithErrorChecking(() => ComplexTypeShortcuts.GetFromShortcut(node.Value, type));
+
+            if (type.IsAbstract || type.IsInterface)
+            {
+                var concreteTypeNode = node.GetChildAddressedByName(KEY_CONCRETE_TYPE);
+                var concreteType = GetNodeData<Type>(concreteTypeNode);
+                return GetNodeData(node, concreteType);
+            }
 
             return RetrieveDataWithErrorChecking(() => ComplexTypes.RetrieveComplexType(node, type));